Separación de la estructura en archivos individuales
