+++ IronMind/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -1,6 +1,7 @@
 {
   "images" : [
     {
+      "filename" : "ChatGPT Image 12 de jun. de 2025 22_58_43.png",
       "idiom" : "universal",
       "platform" : "ios",
       "size" : "1024x1024"
